wip: start command refactoring to support root available units table
