+++ src/MakaMek.Avalonia/MakaMek.Avalonia/Views/StartNewGame/Fragments/PlayersFragment.axaml
@@ -122,7 +122,6 @@
         <Flyout>
             <templatedControls:GamePanel
                 Grid.Row="1"
-                IsVisible="{Binding IsTableVisible}"
                 Title="Select Target Location"
                 CloseCommand="{Binding HideTableCommand}"
                 HorizontalAlignment="Right"
